feat: enhance time and view formatting utilities for improved readability
